+++ extensions/tidyread---streamline-your-daily-reading/src/utils/request.ts
@@ -19,6 +19,37 @@ export async function request(url: URL | string, options?: RequestInit, timeout?
   );
 }
 
+export async function fetchHeadContent(url: string): Promise<string | null> {
+  const response = await request(url);
+
+  // 确保响应是文本类型
+  if (!response.headers.get("content-type")?.includes("text/html")) {
+    console.log("Response is not HTML text");
+    return null;
+  }
+
+  return new Promise((resolve, reject) => {
+    let content = "";
+    response.body?.on?.("data", (chunk: Buffer) => {
+      content += chunk.toString();
+      // 检查是否包含 </head>
+      if (content.includes("</head>")) {
+        response.body!.pause(); // 停止接收数据
+        resolve(content.split("</head>")[0] + "</head>");
+      }
+    });
+
+    response.body?.on?.("error", (err: Error) => {
+      reject(err);
+    });
+
+    response.body?.on?.("end", () => {
+      // 如果没有找到 </head>，则返回已接收的全部内容
+      resolve(content);
+    });
+  });
+}
+
 interface Metadata {
   title: string;
   favicon: string;
@@ -31,9 +62,11 @@ interface Metadata {
 }
 
 export async function fetchMetadata(url: string): Promise<Metadata> {
-  const response = await request(url);
-  const body = await response.text();
-  const $ = load(body);
+  // const response = await request(url);
+  // 如果用这种方式，当文本内容过大时（比如2M），会导致内存溢出，导致页面崩溃
+  // const body = await response.text();
+  const body = await fetchHeadContent(url);
+  const $ = load(body || "");
 
   const title = $("title").text();
   let favicon = $('link[rel="icon"], link[rel="shortcut icon"]').attr("href") ?? "";