+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -6,6 +6,7 @@
 - improve error handling and error message
 - add Delete All Sources
 - add doc for two commands
+- fix fetch large html content memory leak
 
 ## [Initial Version] - 2024-01-06
 init